Relax -—skip & -—until mm:ss.ss specifications to always include "."
In interpreting the mm:ss.ss 'skip' & 'until' specifications, flac
makes use of 'strtod', which applies the conventions of the locale
currently in force. Which means accepting one of '.' or ',' as the
decimal separator. Which one depends on the locale currently in force.

If the separator used on the flac command line does not match the
separator specified by the locale, flac will simply fail, with a message
of the form 'ERROR: invalid value for skip'.

Logitech Media Server makes use of the 'skip' and 'until' specifiers in
in some of its transcoding rules. But it uses the '.' convention
exclusively, without reference to the current locale.
'Slim::Utils::DateTime::fracSecToMinSec' is responsible for formatting
the specification.

In consequence, transcoding silently fails on any platform that is not
set up to use '.' as its decimal separator. Which covers much of the
world.

This change adjusts flac's behaviour such that it uses the "C" locale to
interpret the mm:ss.ss specification. If that fails, it falls back to
the currently active locale.

'strtod_l' is used to make the interpretation under the "C" locale.
Setting up the necessary locale infrastructure is complicated by
variations in the approach taken by different platforms.